+++ src/Symfony/Component/Security/Core/Tests/Authorization/AuthorizationCheckerTest.php
@@ -42,6 +42,9 @@ class AuthorizationCheckerTest extends TestCase
         );
     }
 
+    /**
+     * @group legacy
+     */
     public function testVoteAuthenticatesTokenIfNecessary()
     {
         $token = new UsernamePasswordToken('username', 'password', 'provider');