+++ src/Symfony/Component/Security/Core/CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
  * Deprecate setting the 4th argument (`$alwaysAuthenticate`) to `true` and not setting the
    5th argument (`$exceptionOnNoToken`) to `false` of `AuthorizationChecker`
+ * Deprecate methods `TokenInterface::isAuthenticated()` and `setAuthenticated`,
+   tokens will always be considered authenticated in 6.0
 
 5.3
 ---